+++ src/http-request.ts
@@ -1,4 +1,4 @@
-import { HttpRequest as UHttpRequest } from 'uWebSockets.js';
+import { HttpRequest as UHttpRequest, HttpResponse as UHttpResponse, getParts } from 'uWebSockets.js';
 import { parseQuery } from './utils';
 
 export interface UploadedFile {
@@ -9,111 +9,134 @@ export interface UploadedFile {
 
 export class HttpRequest {
   /**
-   * Request body content
-   */
-  body: { [key: string]: any } = {};
-
-  /**
    * Request user defined data
    */
   data: { [key: string]: any } = {};
 
+  private bodyData: Buffer | null = null;
+  private pattern: string;
+  private req: UHttpRequest;
+  private res: UHttpResponse;
+
+  constructor(req: UHttpRequest, res: UHttpResponse, pattern: string) {
+    this.req = req;
+    this.res = res;
+    this.pattern = pattern;
+  }
+
   /**
    * Request body content
    */
-  files: { [key: string]: UploadedFile | undefined } = {};
+  async body(): Promise<{ [key: string]: any }> {
+    const contentType = this.req.getHeader('content-type');
+
+    if (!contentType) return {};
+
+    const body = this.bodyData ? this.bodyData : await this.getBody(this.res);
+
+    if (!body) return {};
+
+    this.bodyData = body;
+
+    if (contentType === 'application/json' || contentType === 'application/x-www-form-urlencoded') {
+      const bodyStr = body.toString();
+      
+      if (!bodyStr) return {};
+
+      return contentType === 'application/json' ? JSON.parse(body.toString()) : parseQuery(body.toString());
+    } else if (contentType?.startsWith('multipart/form-data')) {
+      const data: any = {};
+
+      getParts(body, contentType)?.forEach(p => {
+        if (!p.type && !p.filename) data[p.name] = Buffer.from(p.data).toString();
+      });
+
+      return data;
+    } else return body;
+  }
+
+  /**
+   * Request body content
+   */
+  async files(): Promise<{ [key: string]: UploadedFile | undefined }> {
+    const contentType = this.req.getHeader('content-type');
+
+    if (!contentType) return {}
+
+    const body = this.bodyData ? this.bodyData : await this.getBody(this.res);
+
+    if (!body) return {};
+
+    this.bodyData = body;
+
+    if (contentType?.startsWith('multipart/form-data')) {
+      const data: any = {};
+
+      getParts(body, contentType)?.forEach(p => {
+        if (p.type && p.filename) data[p.name] = { data: p.data, filename: p.filename, type: p.type };
+      });
+
+      return data;
+    } else return {};
+  }
 
   /**
    * Request headers
    */
-  headers: { [key: string]: string } = {};
+  headers(): { [key: string]: string } {
+    const headers: any = {};
+
+    this.req.forEach((key, value) => headers[key] = value);
+
+    return headers;
+  }
 
   /**
    * Request path params
    */
-  params: { [key: string]: string } = {};
+  params(): { [key: string]: string } {
+    const params = this.pattern.match(/:[\w]+/g);
+
+    if (!params) return {};
+
+    const data: any = {};
+
+    for (let i = 0; i < params.length; i++) data[params[i].replace(':', '')] = this.req.getParameter(i);
+
+    return data;
+  }
 
   /**
    * Request query params
    */
-  query: any = {};
+  query(): { [key: string]: any } {
+    const query = this.req.getQuery();
 
-  /**
-   * Request query params and body content
-   */
-  inputs: { [key: string]: any } = {};
+    if (query) return parseQuery(query);
+
+    return {};
+  }
 
   /**
    * Request URL including initial /slash
    */
-  url: string;
-
-  private pattern: string;
-  private req: UHttpRequest;
-
-  constructor(req: UHttpRequest, pattern: string) {
-    this.req = req;
-    this.pattern = pattern;
-    this.url = req.getUrl();
-
-    this.getHeaders();
-    this.getParams();
-    this.getQuery();
-    this.getInputs();
+  url() {
+    return this.req.getUrl();
   }
 
-  private getHeaders() {
-    this.req.forEach((key, value) => this.headers[key] = value);
-  }
+  private async getBody(res: UHttpResponse): Promise<Buffer> {
+    let buffer: Buffer;
 
-  private getParams() {
-    const params = this.pattern.match(/:[\w]+/g);
+    return new Promise(resolve => res.onData((ab, isLast) => {
+      const chunk = Buffer.from(ab);
 
-    if (!params) return;
-
-    for (let i = 0; i < params.length; i++) this.params[params[i].replace(':', '')] = this.req.getParameter(i);
-  }
-
-  private getQuery() {
-    const query = this.req.getQuery();
-
-    if (query) this.query = parseQuery(query);
-  }
-
-  private getInputs() {
-    this.inputs = { ...this.query, ...this.body };
-  }
-
-  public post(key: string, defaultValue?: any) {
-    return this.body[key] ?? defaultValue ?? null;
-  }
-
-  public all() {
-    this.getInputs();
-    return this.inputs;
-  }
-  
-  public get(key: string, defaultValue?: any) {
-    return this.inputs[key] ?? defaultValue ?? null;
-  }
-
-  public has(key: string) {
-    return this.inputs[key] !== undefined;
-  }
-
-  public header(key: string, defaultValue?: string) {
-    return this.headers[key] ?? defaultValue ?? null;
-  }
-  
-  public param(key: string, defaultValue?: string) {
-    return this.params[key] ?? defaultValue ?? null;
-  }
-
-  public queryParam(key: string, defaultValue?: string) {
-    return this.query[key] ?? defaultValue ?? null;
-  }
-
-  public file(key: string) {
-    return this.files[key] ?? null;
+      if (isLast) {
+        if (buffer) resolve(Buffer.concat([buffer, chunk]));
+        else resolve(chunk);
+      } else {
+        if (buffer) buffer = Buffer.concat([buffer, chunk]);
+        else buffer = Buffer.concat([chunk]);
+      }
+    }));
   }
 }