+++ CHANGELOG.md
@@ -1,9 +1,43 @@
+# [0.4.0](https://github.com/ionited/fiber/compare/0.3.0...0.4.0) (2023-10-25)
+
+### Features
+
+* improve performance getting props on demand ([fbe7a34](https://github.com/ionited/fiber/commit/fbe7a3496b28639e1e32cc67405f3dd642fbcf6a))
+
+### BREAKING CHANGES
+
+* HttpRequest body, files, headers, params, query and url is now functions
+
+# [0.3.0](https://github.com/ionited/fiber/compare/0.2.2...0.3.0) (2023-10-17)
+
+### Bug Fixes
+
+* only parse params if it exists ([7d92c58](https://github.com/ionited/fiber/commit/7d92c58aa68b5bab816cbd16fce13809060b10cd))
+
+### Features
+
+* general simplifications ([f5ea979](https://github.com/ionited/fiber/commit/f5ea9799d1bd7797624c59284cbe2f23628b14eb))
+* notFound handler ([a9b3eb4](https://github.com/ionited/fiber/commit/a9b3eb44bdd38b2e4b28fca82f8e37642ab9a09b))
+
+## [0.2.3](https://github.com/ionited/fiber/compare/0.2.1...0.2.3) (2023-10-13)
+
+### Features
+
+* general simplifications ([f5ea979](https://github.com/ionited/fiber/commit/f5ea9799d1bd7797624c59284cbe2f23628b14eb))
+
+## [0.2.2](https://github.com/ionited/fiber/compare/0.2.1...0.2.2) (2023-10-05)
+
+## [0.2.1](https://github.com/ionited/fiber/compare/0.2.0...0.2.1) (2023-09-13)
+
+### Bug Fixes
+
+* parse array values on query ([9188ed6](https://github.com/ionited/fiber/commit/9188ed6dd42226f1a5b2558b076a13258e6ae162))
+
 # 0.2.0 (2023-07-19)
 
 ### Bug Fixes
 
 * missing function pattern on documentation ([f67edba](https://github.com/ionited/fiber/commit/f67edba65cc1c0eef596c430eb736d6322afd585))
-
 
 ### Features